+++ src/entities/book/index.ts
@@ -1,1 +1,2 @@
 export { default as BookCard } from './ui/BookCard';
+export { default as BookList } from './ui/BookList';
