+++ src/entities/book/model/book.ts
@@ -1,0 +1,8 @@
+export interface Book {
+  id: number;
+  title: string;
+  author: string;
+  publisher: string;
+  publishDate: string;
+  coverImage: string;
+}
