+++ src/entities/book/ui/BookList.tsx
@@ -1,0 +1,19 @@
+import { Book } from '../model/book';
+import BookCard from './BookCard';
+
+interface BookListProps {
+  books: Book[];
+}
+
+export default function BookList({ books }: BookListProps) {
+  return (
+    <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4">
+      {books.map((book) => (
+        <BookCard
+          key={book.id}
+          book={book}
+        />
+      ))}
+    </div>
+  );
+}
