+++ src/features/book-filter/index.ts
@@ -1,0 +1,2 @@
+export { default as BookFilter } from './ui/BookFilter';
+export type { FilterType } from './model/filter-type';
