+++ src/features/book-filter/model/filter-type.ts
@@ -1,0 +1,1 @@
+export type FilterType = 'new' | 'bestseller';
