+++ src/features/book-search/index.ts
@@ -1,0 +1,1 @@
+export { default as BookSearch } from './ui/BookSearch';
